try this (#1428)
* try this

* speed things up with update demographic codes

* add VCR for mini bundle

* use mini bundle vcr

* remove code descriptions from API measure evaluator
