+++ lib/cypress/api_measure_evaluator.rb
@@ -510,7 +510,6 @@ module Cypress
 
     def import_cat1_file(doc, patient_ids, bundle_id)
       patient, _warnings, codes = QRDA::Cat1::PatientImporter.instance.parse_cat1(doc)
-      Cypress::QRDAPostProcessor.build_code_descriptions(codes, patient, bundle)
       Cypress::QRDAPostProcessor.replace_negated_codes(patient, Bundle.find(bundle_id))
       patient.update(_type: CQM::TestExecutionPatient, correlation_id: 'api_eval')
       patient.save!
